requirements.txt
docopts
python-dotenv
requests
PyMsTeamsAdaptiveCard